Release Version of the Layout
Been working on this 2 weeks and the first working version is finished. Was hard to understand the docs, mainly because my mother language isn't english but I did my best. It is a personal project for Haloruns.com and I know it is not the best wat to use nodecg but the program does what it is required.